requirements.txt
# Core
python-dotenv==1.0.0
requests
pandas
numpy
tqdm

# Database
pymongo==4.6.1
SQLAlchemy

# Text Cleaning & NLP
regex
nltk
emoji
langdetect
langid

# HuggingFace & Training
transformers>=4.2.0
datasets
sentencepiece
torch
scikit-learn

# Model Serving
fastapi
uvicorn
flask

# Dashboard (Step 4)
streamlit

# Utilities
joblib
Faker
google-generativeai